Code update on google to latest revision as on 12th Feb 2013. (revision 4338)

+++ trunk/dcma-gwt/dcma-gwt-admin/src/main/java/com/ephesoft/dcma/gwt/admin/bm/server/ImportBatchClassUploadServlet.java
@@ -1,6 +1,6 @@
 /********************************************************************************* 
 * Ephesoft is a Intelligent Document Capture and Mailroom Automation program 
-* developed by Ephesoft, Inc. Copyright (C) 2010-2011 Ephesoft Inc. 
+* developed by Ephesoft, Inc. Copyright (C) 2010-2012 Ephesoft Inc. 
 * 
 * This program is free software; you can redistribute it and/or modify it under 
 * the terms of the GNU Affero General Public License version 3 as published by the 
@@ -54,58 +54,79 @@ import org.apache.commons.fileupload.FileUploadException;
 import org.apache.commons.fileupload.disk.DiskFileItemFactory;
 import org.apache.commons.fileupload.servlet.ServletFileUpload;
 import org.apache.commons.io.FilenameUtils;
+import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.SerializationUtils;
-
 
 import com.ephesoft.dcma.batch.service.BatchSchemaService;
 import com.ephesoft.dcma.batch.service.ImportBatchService;
 import com.ephesoft.dcma.core.common.FileType;
 import com.ephesoft.dcma.da.domain.BatchClass;
 import com.ephesoft.dcma.da.service.BatchClassService;
+import com.ephesoft.dcma.gwt.admin.bm.client.i18n.BatchClassManagementConstants;
 import com.ephesoft.dcma.gwt.core.server.DCMAHttpServlet;
 import com.ephesoft.dcma.util.FileUtils;
 import com.ephesoft.dcma.workflow.service.common.DeploymentService;
 
+/**
+ * This is class for importing batch class upload Servlet.
+ * 
+ * @author Ephesoft
+ * @version 1.0
+ * @see com.ephesoft.dcma.gwt.core.server.DCMAHttpServlet
+ */
 public class ImportBatchClassUploadServlet extends DCMAHttpServlet {
 
 	/**
-	 * 
+	 * serialVersionUID long.
 	 */
 	private static final long serialVersionUID = 1L;
 
+	/**
+	 * SERIALIZATION_EXT String.
+	 */
 	private static final String SERIALIZATION_EXT = FileType.SER.getExtensionWithDot();
 
+	/**
+	 * Overridden doGet method.
+	 * 
+	 * @param request HttpServletRequest
+	 * @param response HttpServletResponse
+	 * @throws IOException
+	 */
 	@Override
 	public final void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
 		doPost(request, response);
 	}
 
+	/**
+	 * Overridden doPost method.
+	 * 
+	 * @param request HttpServletRequest
+	 * @param response HttpServletResponse
+	 * @throws IOException
+	 */
 	@Override
 	public void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {
 		BatchSchemaService batchSchemaService = this.getSingleBeanOfType(BatchSchemaService.class);
 		DeploymentService deploymentService = this.getSingleBeanOfType(DeploymentService.class);
 		BatchClassService bcService = this.getSingleBeanOfType(BatchClassService.class);
 		ImportBatchService imService = this.getSingleBeanOfType(ImportBatchService.class);
-		
+
 		String lastAttachedZipSourcePath = req.getParameter("lastAttachedZipSourcePath");
-		if (lastAttachedZipSourcePath != null && !lastAttachedZipSourcePath.isEmpty()) {
-			if (new File(lastAttachedZipSourcePath).exists()) {
-				FileUtils.deleteDirectoryAndContentsRecursive(new File(lastAttachedZipSourcePath));
-			}
+		if (lastAttachedZipSourcePath != null && !lastAttachedZipSourcePath.isEmpty() && new File(lastAttachedZipSourcePath).exists()) {
+			FileUtils.deleteDirectoryAndContentsRecursive(new File(lastAttachedZipSourcePath));
 		}
 		attachFile(req, resp, batchSchemaService, deploymentService, bcService, imService);
 	}
 
-	private void attachFile(HttpServletRequest req, HttpServletResponse resp, BatchSchemaService batchSchemaService, DeploymentService deploymentService, BatchClassService bcService, ImportBatchService imService)
-			throws IOException {
-	
-		
+	private void attachFile(HttpServletRequest req, HttpServletResponse resp, BatchSchemaService batchSchemaService,
+			DeploymentService deploymentService, BatchClassService bcService, ImportBatchService imService) throws IOException {
 		PrintWriter printWriter = resp.getWriter();
 		File tempZipFile = null;
 		InputStream instream = null;
 		OutputStream out = null;
-		String zipWorkFlowName = "", tempOutputUnZipDir = "";
-		BatchClass importBatchClass = null; 
+		String zipWorkFlowName = BatchClassManagementConstants.EMPTY_STRING, tempOutputUnZipDir = BatchClassManagementConstants.EMPTY_STRING, systemFolderPath = BatchClassManagementConstants.EMPTY_STRING;
+		BatchClass importBatchClass = null;
 		if (ServletFileUpload.isMultipartContent(req)) {
 			FileItemFactory factory = new DiskFileItemFactory();
 			ServletFileUpload upload = new ServletFileUpload(factory);
@@ -114,22 +135,18 @@ public class ImportBatchClassUploadServlet extends DCMAHttpServlet {
 			if (!exportSerailizationFolder.exists()) {
 				exportSerailizationFolder.mkdir();
 			}
-
-			String zipFileName = "";
-			String zipPathname = "";
+			String zipFileName = BatchClassManagementConstants.EMPTY_STRING;
+			String zipPathname = BatchClassManagementConstants.EMPTY_STRING;
 			List<FileItem> items;
-			
 			try {
-				items = upload.parseRequest(req);
+				items = (List<FileItem>) upload.parseRequest(req);
 				for (FileItem item : items) {
-
 					if (!item.isFormField() && "importFile".equals(item.getFieldName())) {
 						zipFileName = item.getName();
 						if (zipFileName != null) {
 							zipFileName = zipFileName.substring(zipFileName.lastIndexOf(File.separator) + 1);
 						}
 						zipPathname = exportSerailizationFolderPath + File.separator + zipFileName;
-						// get only the file name not whole path
 						if (zipFileName != null) {
 							zipFileName = FilenameUtils.getName(zipFileName);
 						}
@@ -140,85 +157,71 @@ public class ImportBatchClassUploadServlet extends DCMAHttpServlet {
 								tempZipFile.delete();
 							}
 							out = new FileOutputStream(tempZipFile);
-							byte buf[] = new byte[1024];
-							int len;
-							while ((len = instream.read(buf)) > 0) {
+							byte buf[] = new byte[BatchClassManagementConstants.BUFFER_SIZE];
+							int len = instream.read(buf);
+							while ((len) > 0) {
 								out.write(buf, 0, len);
+								len = instream.read(buf);
 							}
 						} catch (FileNotFoundException e) {
-							log.error("Unable to create the export folder." + e, e);
+							LOG.error("Unable to create the export folder." + e, e);
 							printWriter.write("Unable to create the export folder.Please try again.");
 
 						} catch (IOException e) {
-							log.error("Unable to read the file." + e, e);
+							LOG.error("Unable to read the file." + e, e);
 							printWriter.write("Unable to read the file.Please try again.");
 						} finally {
-							if (out != null) {
-								try {
-									out.close();
-								} catch (IOException ioe) {
-									log.info("Could not close stream for file." + tempZipFile);
-								}
-							}
-							if (instream != null) {
-								try {
-									instream.close();
-								} catch (IOException ioe) {
-									log.info("Could not close stream for file." + zipFileName);
-								}
-							}
+							IOUtils.closeQuietly(out);
+							IOUtils.closeQuietly(instream);
 						}
 					}
 				}
 			} catch (FileUploadException e) {
-				log.error("Unable to read the form contents." + e, e);
+				LOG.error("Unable to read the form contents." + e, e);
 				printWriter.write("Unable to read the form contents.Please try again.");
 			}
-
 			tempOutputUnZipDir = exportSerailizationFolderPath + File.separator
 					+ zipFileName.substring(0, zipFileName.lastIndexOf('.'));
 			try {
 				FileUtils.unzip(tempZipFile, tempOutputUnZipDir);
 			} catch (Exception e) {
-				log.error("Unable to unzip the file." + e, e);
+				LOG.error("Unable to unzip the file." + e, e);
 				printWriter.write("Unable to unzip the file.Please try again.");
 				tempZipFile.delete();
 			}
-
 			String serializableFilePath = FileUtils.getFileNameOfTypeFromFolder(tempOutputUnZipDir, SERIALIZATION_EXT);
 			InputStream serializableFileStream = null;
-			
-				
 			try {
 				serializableFileStream = new FileInputStream(serializableFilePath);
 				importBatchClass = (BatchClass) SerializationUtils.deserialize(serializableFileStream);
 				zipWorkFlowName = importBatchClass.getName();
+				systemFolderPath = importBatchClass.getSystemFolder();
+				if (systemFolderPath == null) {
+					systemFolderPath = BatchClassManagementConstants.EMPTY_STRING;
+				}
 			} catch (Exception e) {
 				tempZipFile.delete();
-				log.error("Error while importing" + e, e);
+				LOG.error("Error while importing" + e, e);
 				printWriter.write("Error while importing.Please try again.");
 			} finally {
-				if (serializableFileStream != null) {
-					try {
-						serializableFileStream.close();
-					} catch (IOException ioe) {
-						log.info("Could not close stream for file." + serializableFilePath);
-					}
-				}
+				IOUtils.closeQuietly(serializableFileStream);
 			}
-
 		} else {
-			log.error("Request contents type is not supported.");
+			LOG.error("Request contents type is not supported.");
 			printWriter.write("Request contents type is not supported.");
 		}
 		if (tempZipFile != null) {
 			tempZipFile.delete();
 		}
-		
-		List<String> uncList = bcService.getAssociatedUNCList(zipWorkFlowName);		
+		List<String> uncList = bcService.getAssociatedUNCList(zipWorkFlowName);
 		boolean isWorkflowDeployed = deploymentService.isDeployed(zipWorkFlowName);
 		boolean isWorkflowEqual = imService.isImportWorkflowEqualDeployedWorkflow(importBatchClass, importBatchClass.getName());
-		
+		printWriterMethod(printWriter, zipWorkFlowName, tempOutputUnZipDir, systemFolderPath, uncList, isWorkflowDeployed,
+				isWorkflowEqual);
+	}
+
+	private void printWriterMethod(PrintWriter printWriter, String zipWorkFlowName, String tempOutputUnZipDir,
+			String systemFolderPath, List<String> uncList, boolean isWorkflowDeployed, boolean isWorkflowEqual) {
 		printWriter.write("workFlowName:" + zipWorkFlowName);
 		printWriter.append("|");
 		printWriter.append("filePath:").append(tempOutputUnZipDir);
@@ -227,9 +230,10 @@ public class ImportBatchClassUploadServlet extends DCMAHttpServlet {
 		printWriter.append("|");
 		printWriter.write("workflowEqual:" + isWorkflowEqual);
 		printWriter.append("|");
-		printWriter.write("workflowExistInBatchClass:" + ((uncList == null || uncList.size() == 0)? false : true));
+		printWriter.write("workflowExistInBatchClass:" + ((uncList == null || uncList.size() == 0) ? false : true));
+		printWriter.append("|");
+		printWriter.write("systemFolderPath:" + systemFolderPath);
 		printWriter.append("|");
 		printWriter.flush();
 	}
 }
-
